refactor(test): rename 20 datastore *IT → *Test — Testcontainers tests now contribute to JaCoCo surefire coverage (59→67% instruction, 53→57% branch)
