+++ src/test/java/ai/labs/eddi/datastore/mongo/MongoAuditStoreTest.java
@@ -1,13 +1,9 @@
-package ai.labs.eddi.datastore.postgres;
+package ai.labs.eddi.datastore.mongo;
 
-import ai.labs.eddi.datastore.serialization.IJsonSerialization;
-import ai.labs.eddi.datastore.serialization.JsonSerialization;
+import ai.labs.eddi.engine.audit.AuditStore;
 import ai.labs.eddi.engine.audit.model.AuditEntry;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.*;
 
-import javax.sql.DataSource;
-import java.sql.SQLException;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
@@ -17,30 +13,23 @@ import java.util.UUID;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
- * Integration tests for {@link PostgresAuditStore} using Testcontainers.
+ * Integration tests for {@link AuditStore} (MongoDB) using Testcontainers.
  *
  * @since 6.0.0
  */
-@DisplayName("PostgresAuditStore IT")
-class PostgresAuditStoreIT extends PostgresTestBase {
+@DisplayName("MongoAuditStore IT")
+class MongoAuditStoreTest extends MongoTestBase {
 
-    private static PostgresAuditStore store;
-    private static DataSource ds;
+    private static AuditStore store;
 
     @BeforeAll
     static void init() {
-        var dsInstance = createDataSourceInstance();
-        ds = dsInstance.get();
-        IJsonSerialization json = new JsonSerialization(new ObjectMapper());
-        store = new PostgresAuditStore(dsInstance, json);
+        store = new AuditStore(getDatabase());
     }
 
     @BeforeEach
     void clean() {
-        try {
-            truncateTables(ds, "audit_ledger");
-        } catch (SQLException ignored) {
-        }
+        dropCollections("audit_ledger");
     }
 
     // ─── Core CRUD ──────────────────────────────────────────────
@@ -63,7 +52,7 @@ class PostgresAuditStoreIT extends PostgresTestBase {
         }
 
         @Test
-        @DisplayName("multiple entries — ordered by created_at DESC")
+        @DisplayName("multiple entries — ordered by timestamp DESC")
         void multipleEntries() {
             store.appendEntry(createEntry("conv1", "agent1", 1, "user1", "parser", "input", 0, 0, 10L));
             store.appendEntry(createEntry("conv1", "agent1", 1, "user1", "behavior", "rules", 0, 1, 20L));
@@ -71,7 +60,6 @@ class PostgresAuditStoreIT extends PostgresTestBase {
 
             List<AuditEntry> results = store.getEntries("conv1", 0, 10);
             assertEquals(3, results.size());
-            // All 3 entries should be present — verify by taskId
             var taskIds = results.stream().map(AuditEntry::taskId).toList();
             assertTrue(taskIds.contains("parser"));
             assertTrue(taskIds.contains("behavior"));
@@ -213,11 +201,11 @@ class PostgresAuditStoreIT extends PostgresTestBase {
     // ─── JSONB data round-trip ──────────────────────────────────
 
     @Test
-    @DisplayName("JSONB data — input/output/actions round-trip")
-    void jsonbDataRoundTrip() {
+    @DisplayName("BSON data — input/output/actions round-trip")
+    void bsonDataRoundTrip() {
         var entry = new AuditEntry(UUID.randomUUID().toString(), "conv_json", "agent1", 1,
                 "user1", "test", 0, "llm_task", "langchain", 0, 100L,
-                Map.of("userInput", "hello"), Map.of("output", List.of("world")),
+                Map.of("userInput", "hello"), Map.of("output", "world"),
                 Map.of("compiledPrompt", "You are...", "modelName", "gpt-4"),
                 null, List.of("greet", "respond"), 0.05,
                 Instant.now().truncatedTo(ChronoUnit.MILLIS), "hmac_val", null);
